Node: assume (but verify) that Buffers aren't modified while in use
A JsBuffer owns its storage*, so it's safe to assume the buffer won't
get deallocated. What's unsafe is assuming that no one else will
modify the buffer while we have a reference to it, which is why we
checksum it. (We can't stop the Rust compiler from potentially
optimizing out that checksum, though.)

To keep from slowing down operations too much, the checksum is limited
to the first 1024 bytes of a buffer unless debug logging is turned on.

* https://nodejs.org/api/n-api.html#n_api_napi_get_buffer_info